refactor(install): group MCP modules under install/mcp/ subpackage
Move the seven mcp_*.py helpers under apm_cli/install/ into a focused
mcp/ subpackage and strip the redundant `mcp_` prefix from each
filename. Aligns with the unprefixed-noun naming convention used by
the rest of install/ (pipeline.py, context.py, service.py, ...).

Pure refactor: behaviour, public symbol names, and the back-compat
re-bound `_xxx` aliases in commands/install.py are unchanged. All
install/ unit tests, the #938 regression, and the LOC budget invariant
stay green.
